requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.4.1
click==8.1.8
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
Pygments==2.19.1
PyYAML==6.0.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
